requirements.txt
customtkinter==5.2.2
Pillow==10.2.0
matplotlib==3.10.3
requests==2.32.3
subprocess
sys
os